pyproject.toml
[tool.poetry]
name = "shin-roulette"
version = "0.1.0"
description = ""
authors = ["Ricimon <ricimon25@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "shin_roulette"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^5.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
yapf = "^0.40.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
